DOP-5071 update types and file structure
