+++ populate-metadata/src/updateConfig.ts
@@ -1,14 +1,63 @@
+import type { NetlifyConfig } from '@netlify/build';
+import type { DbConfig, Environments } from './assertDbEnvVars';
+import type { DocsetsDocument } from './databaseConnection/fetchDocsetsData';
+import type {
+  ReposBranchesDocument,
+  BranchEntry,
+} from './databaseConnection/fetchReposBranchesData';
 import { getProperties } from './getProperties';
-import type { DocsConfig, DbConfig, Environments } from './types';
+
+export interface DocsConfig extends Omit<NetlifyConfig, 'build'> {
+  build: Build;
+}
+
+type ConfigEnvironmentVariables = Partial<{
+  BRANCH: string;
+  SITE_NAME: string;
+  INCOMING_HOOK_URL: string;
+  INCOMING_HOOK_TITLE: string;
+  INCOMING_HOOK_BODY: string;
+  ENV: Environments;
+  REPO_ENTRY: ReposBranchesDocument;
+  DOCSET_ENTRY: DocsetsDocument;
+  BRANCH_ENTRY: BranchEntry[];
+}>;
+
+export interface Build {
+  command?: string;
+  publish: string;
+  base: string;
+  services: Record<string, unknown>;
+  ignore?: string;
+  edge_handlers?: string;
+  edge_functions?: string;
+  environment: ConfigEnvironmentVariables;
+  processing: {
+    skip_processing?: boolean;
+    css: {
+      bundle?: boolean;
+      minify?: boolean;
+    };
+    js: {
+      bundle?: boolean;
+      minify?: boolean;
+    };
+    html: {
+      pretty_url?: boolean;
+    };
+    images: {
+      compress?: boolean;
+    };
+  };
+}
 
 export const updateConfig = async (
-  config: DocsConfig,
+  //TODO: make this more specific to just environment
+  configEnvironment: ConfigEnvironmentVariables,
   dbEnvVars: DbConfig,
 ): Promise<void> => {
-  const branchName =
-    process.env.BRANCH_NAME ?? config.build?.environment?.BRANCH;
-  const repoName =
-    process.env.REPO_NAME ?? config.build?.environment?.SITE_NAME;
+  const branchName = process.env.BRANCH_NAME ?? configEnvironment.BRANCH;
+  const repoName = process.env.REPO_NAME ?? configEnvironment.SITE_NAME;
 
   if (!branchName || !repoName) {
     throw new Error('Repo name or branch name missing from deploy');
@@ -17,15 +66,15 @@ export const updateConfig = async (
   // Check if build was triggered by a webhook
   // TODO: add more specific logic dependent on hook title, url, body, etc. once Slack deploy apps have been implemented
   const isWebhookDeploy = !!(
-    config.build.environment?.INCOMING_HOOK_URL &&
-    config.build.environment?.INCOMING_HOOK_TITLE &&
-    config.build.environment?.INCOMING_HOOK_BODY
+    configEnvironment.INCOMING_HOOK_URL &&
+    configEnvironment.INCOMING_HOOK_TITLE &&
+    configEnvironment.INCOMING_HOOK_BODY
   );
 
   // Check if this was an engineering build or writer's build; writer's builds by default are all builds not built on the "mongodb-snooty" site
   // Environment is either dotcomprd or prd if it is a writer build
-  config.build.environment.ENV =
-    config.build.environment.SITE_NAME === 'mongodb-snooty'
+  configEnvironment.ENV =
+    configEnvironment.SITE_NAME === 'mongodb-snooty'
       ? isWebhookDeploy
         ? 'dotcomstg'
         : 'stg'
@@ -37,22 +86,22 @@ export const updateConfig = async (
     branchName: branchName,
     repoName: repoName,
     dbEnvVars: dbEnvVars,
-    environment: config.build.environment.ENV,
+    environment: configEnvironment.ENV,
   });
 
   const { branches: branch, ...repoEntry } = repo;
-  config.build.environment.REPO_ENTRY = repoEntry;
-  config.build.environment.DOCSET_ENTRY = docsetEntry;
-  config.build.environment.BRANCH_ENTRY = branch;
+  configEnvironment.REPO_ENTRY = repoEntry;
+  configEnvironment.DOCSET_ENTRY = docsetEntry;
+  configEnvironment.BRANCH_ENTRY = branch;
 
   console.info(
     'BUILD ENVIRONMENT: ',
-    config.build.environment.ENV,
+    configEnvironment.ENV,
     '\n REPO ENTRY: ',
-    config.build.environment.REPO_ENTRY,
+    configEnvironment.REPO_ENTRY,
     '\n DOCSET ENTRY: ',
-    config.build.environment.DOCSET_ENTRY,
+    configEnvironment.DOCSET_ENTRY,
     '\n BRANCH ENTRY: ',
-    config.build.environment.BRANCH_ENTRY,
+    configEnvironment.BRANCH_ENTRY,
   );
 };
